feat(dashboard): separate fastest performers into two cards
